requirements.txt
openai
python-dotenv
instructor
aiohttp
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
fire==0.5.0
frozenlist
idna==3.4
iniconfig==2.0.0
multidict
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-mock==3.10.0
python-dotenv==1.0.0
requests==2.29.0
ruff==0.0.263
six==1.16.0
termcolor==2.3.0
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.15
yarl==1.9.2
hamcrest
